Get rid of the extension package
That logic lives in the state package now
