--- extension/package.json
@@ -1,5 +1,0 @@
-{
-  "main": "dist/index.js",
-  "types": "dist/extension.d.ts",
-  "module": "dist/index.es.js"
-}
