--- extension/src/README.md
@@ -1,13 +1,0 @@
-@Extension
-
-@Behavior
-
-@ExtensionGroup
-
-@Configuration
-
-### Helpers
-
-@combineConfig
-
-@fillConfig
