+++ state/src/index.ts
@@ -5,3 +5,4 @@ export {EditorSelection, SelectionRange} from "./selection"
 export {Change, ChangeDesc, ChangeSet, Mapping, MapMode, ChangedRange} from "./change"
 export {Transaction} from "./transaction"
 export {Text} from "../../text"
+export {combineConfig, fillConfig} from "./config"
